% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plug-in.function.R
\name{cluster.lr.dens}
\alias{cluster.lr.dens}
\title{To get interaction density between clusters and receptors}
\usage{
cluster.lr.dens(top.rep.name, object, select.ident, ident.label, find)
}
\description{
To get interaction density between clusters and receptors
}
