% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plug-in.function.R
\name{cluster.lr.inten}
\alias{cluster.lr.inten}
\title{To get interaction intensity between clusters and receptors}
\usage{
cluster.lr.inten(top.rep.name, object, select.ident, ident.label, find)
}
\arguments{
\item{top.rep.name}{top.rep.name}

\item{object}{CommPath object}

\item{select.ident}{select.ident}

\item{ident.label}{ident.label}

\item{find}{find}
}
\value{
matrix of clusters * genes, elements are interaction intensity between genes and clusters
}
\description{
To get interaction intensity between clusters and receptors
}
