% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Commpath-object.R
\docType{class}
\name{Commpath-class}
\alias{Commpath-class}
\alias{Commpath}
\title{To define the Commpath object and the key slots}
\description{
To define the Commpath object and the key slots
}
\section{Slots}{

\describe{
\item{\code{data}}{normalized expression matrix}

\item{\code{meta.info}}{vector of cell labels}

\item{\code{LR.marker}}{data.frame of the differential expression test result of ligands and receptors}

\item{\code{interact}}{list containing information of LR interaction among clusters}

\item{\code{pathway}}{list containing information of pathways associated with ligands and receptors}
}}

