% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CommPathData}
\alias{CommPathData}
\title{Ligand-receptor pairs and pathway annotations in CommPath database}
\format{
A list includes ligand-receptor pairs and pathway annotations including KEGG pathways, WikiPathways, reactome pathways, and GO terms
}
\usage{
CommPathData
}
\description{
The ligand-receptor pairs and pathway annotations curated in CommPath
}
\keyword{datasets}
