% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Commpath.R
\name{compCommpathMarker}
\alias{compCommpathMarker}
\title{To conduct defferential expression test in select.ident between two Commpath objects}
\usage{
compCommpathMarker(
  object,
  select.ident,
  compare,
  method = "wilcox.test",
  p.adjust = "BH",
  only.posi = FALSE,
  only.sig = TRUE
)
}
\arguments{
\item{object}{The first Commpath object}

\item{select.ident}{Identity class of interest}

\item{compare}{compare}

\item{method}{Method used for differential expression test, either 'wilcox.test' or 't.test'}

\item{p.adjust}{Method used for p value correction for multiple differential expression test; see p.adjust function for more information}

\item{only.posi}{only logFC > 0}

\item{only.sig}{only p_val_adj < 0.05}
}
\value{
data.frame of differentially expressed geens between the same clusters in two Commpath object
}
\description{
To conduct defferential expression test in select.ident between two Commpath objects
}
