% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commpath.R
\name{diffAllPath}
\alias{diffAllPath}
\title{To find different enriched pathways in each identity class}
\usage{
diffAllPath(Interact, gsva.mat, ident.label, method = "t.test")
}
\arguments{
\item{Interact}{Interact list returned by findLRpath}

\item{gsva.mat}{Matrix containing the pathway enrichment sorces, with rows representing pathways and columns representing cells. Pathway scores are usually computed from gsva, or other methods aiming to measure the pathway enrichment in cells}

\item{ident.label}{Vector indicating the identity labels of cells, and the order of labels are required to match order of cells (columns) in the gsva.mat}

\item{method}{Method used for differential enrichment analysis, either 't.test' of 'wilcox.test'}
}
\value{
Dataframe including the statistic result comparing the pathway enrichment sorces between cells in each cluster and all other clusters, the significant recetor and ligand in the pathways, and the corresponding up stream identity class and ligand
}
\description{
To find different enriched pathways in each identity class
}
