% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Commpath.R
\name{diffCommpath}
\alias{diffCommpath}
\title{To compare two Commpath object}
\usage{
diffCommpath(object.1, object.2, select.ident)
}
\arguments{
\item{object.1}{The first Commpath object}

\item{object.2}{The second Commpath object for comparison}

\item{select.ident}{Identity class of interest}
}
\value{
Network plot showing difference between two Commpath object
}
\description{
To compare two Commpath object
}
