% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommPath.R
\name{diffCommPathMarker}
\alias{diffCommPathMarker}
\title{To conduct defferential expression test in select.ident between two CommPath objects}
\usage{
diffCommPathMarker(
  object.1,
  object.2,
  select.ident,
  method = "wilcox.test",
  p.adjust = "BH",
  only.posi = FALSE,
  only.sig = TRUE
)
}
\arguments{
\item{object.1}{CommPath object 1}

\item{object.2}{CommPath object 2 for comparison}

\item{select.ident}{Identity class of interest}

\item{method}{Method used for differential expression test, either 'wilcox.test' or 't.test'}

\item{p.adjust}{Method used for p value correction for multiple differential expression test; see p.adjust function for more information}

\item{only.posi}{only logFC > 0}

\item{only.sig}{only p_val_adj < 0.05}
}
\value{
Data frame of differentially expressed geens between the same clusters in two CommPath object
}
\description{
To conduct defferential expression test in select.ident between two CommPath objects
}
