% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Commpath.R
\name{diffCommpathPlot}
\alias{diffCommpathPlot}
\title{To compare two Commpath objects}
\usage{
diffCommpathPlot(
  object.1,
  object.2,
  select.ident,
  diff.obj.marker = NULL,
  top.n.receptor = 5,
  order = NULL,
  top.n.path = 10,
  p.thre = 0.05,
  dot.ident.col = NULL,
  dot.receptor.col = NULL,
  bar.pathway.col = NULL,
  bar.pathway.width = 10,
  dot.ident.size = 1,
  dot.receptor.size = 1,
  label.text.size = 1,
  label.title.size = 1,
  line.ident.width = 1,
  line.path.width = 1
)
}
\arguments{
\item{object.1}{The first Commpath object}

\item{object.2}{The second Commpath object for comparison}

\item{select.ident}{Identity class of interest}

\item{diff.obj.marker}{diff.obj.marker computed by diffCommpathMarker}
}
\value{
data.frame of differentially expressed geens between the same clusters in two Commpath object
}
\description{
To compare two Commpath objects
}
