% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommPath-visualization.R
\name{DotPlot.pathway}
\alias{DotPlot.pathway}
\title{To present a dot plot for top ligand-receptor pairs involved in the specific pathways in the selected clusters}
\usage{
DotPlot.pathway(
  object,
  acti.path.dat,
  pathway,
  ligand.ident = NULL,
  receptor.ident = NULL,
  ident.levels = NULL,
  top.n.inter = 10,
  return.data = FALSE
)
}
\arguments{
\item{object}{CommPath object}

\item{acti.path.dat}{Data frame of differential activation test result of filtered from diffAllPath}

\item{pathway}{To present a dot plot for ligand-receptor pairs involved in which pathway}

\item{ligand.ident}{Vector containing the ligand ident}

\item{receptor.ident}{Vector containing the receptor ident}

\item{ident.levels}{Vector of levels of the identities}

\item{top.n.inter}{Show the dotplot for the top n LR pairs with the largest product of Log2FC}

\item{return.data}{Logical value indicating whether to return the data for the plot or not}
}
\value{
Dotplot showing the ligand-receptor interaction involved in the specific pathways in the selected clusters
}
\description{
To present a dot plot for top ligand-receptor pairs involved in the specific pathways in the selected clusters
}
