% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Commpath-visualization.R
\name{dotPlot}
\alias{dotPlot}
\title{To present a dot plot for specific ligand-receptor pairs in specific clusters}
\usage{
dotPlot(
  object,
  ligand.ident = NULL,
  receptor.ident = NULL,
  ident.levels = NULL,
  top.n.inter = 10,
  return.data = FALSE
)
}
\arguments{
\item{object}{Commpath object}

\item{ligand.ident}{Vector containing the ligand ident}

\item{receptor.ident}{Vector containing the receptor ident}

\item{ident.levels}{Vector of levels of the identities}

\item{top.n.inter}{Show the dotplot for the top n LR pairs with the largest product of Log2FC}

\item{return.data}{Logical value indicating whether to return the data for the plot or not}
}
\value{
Dotplot showing the ligand-receptor interaction between the selected ligand.ident and receptor.ident
}
\description{
To present a dot plot for specific ligand-receptor pairs in specific clusters
}
