% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plug-in.function.R
\name{extract.info}
\alias{extract.info}
\title{To extract information from ident.path.dat}
\usage{
extract.info(ident.path.dat)
}
\arguments{
\item{ident.path.dat}{ident.path.dat}
}
\value{
data.frame
}
\description{
To extract information from ident.path.dat
}
