% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plug-in.function.R
\name{factor.to.character}
\alias{factor.to.character}
\title{To convert the variables in a data.frame from factor to character}
\usage{
factor.to.character(x, column = c(1, 2))
}
\arguments{
\item{x}{data.frame with columns named as Cell.From and Cell.To}

\item{column}{which column}
}
\value{
data.frame with variables converted to character
}
\description{
To convert the variables in a data.frame from factor to character
}
