% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommPath.R
\name{filterInter}
\alias{filterInter}
\title{To screen LR pairs involved in activated pathways}
\usage{
filterInter(object, acti.path.dat)
}
\arguments{
\item{object}{CommPath object}

\item{acti.path.dat}{Data frame of differential enrichment test result from diffAllPath}
}
\value{
CommPath object with LR interactions filtered by activated pathways in each cluster
}
\description{
To screen LR pairs involved in activated pathways
}
