% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommPath-main.R
\name{filterLR.path.contain.lig}
\alias{filterLR.path.contain.lig}
\title{To retrieve activated pathways containing the ligand}
\usage{
filterLR.path.contain.lig(object, acti.path.filtered.dat)
}
\arguments{
\item{object}{CommPath object}

\item{acti.path.filtered.dat}{Data frame of differential enrichment test result from diffAllPath}
}
\value{
CommPath object with LR interactions filtered by activated pathways in each cluster
}
\description{
To retrieve activated pathways containing the ligand
}
