% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommPath.R
\name{filterPath}
\alias{filterPath}
\title{To remove those pathways containing only ligands which do not triger any pathway in the downstream clusters}
\usage{
filterPath(object, acti.path.dat)
}
\arguments{
\item{object}{CommPath object}

\item{acti.path.dat}{Data frame of differential enrichment test result from diffAllPath}
}
\value{
Data frame including the statistic result of filtered pathways in each cluster
}
\description{
To remove those pathways containing only ligands which do not triger any pathway in the downstream clusters
}
