% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commpath.R
\name{findLRmarker}
\alias{findLRmarker}
\title{To identify marker ligands and marker receptors in the expression matrix}
\usage{
findLRmarker(expr.mat, label, species, method = "wilcox.test", p.adjust = "BH")
}
\arguments{
\item{expr.mat}{Matrix or data frame of expression matrix, with genes in rows and cells in columns}

\item{label}{Vector of identity classes of cells in the expression matrix}

\item{species}{Species, either 'hsapiens', 'mmusculus', or 'rnorvegicus'}

\item{method}{Method used for differential expression test, either 'wilcox.test' or 't.test'}

\item{p.adjust}{Method used for p value correction for multiple differential expression test; see p.adjust function for more information}
}
\value{
Data frame containing the differential expression test
}
\description{
To identify marker ligands and marker receptors in the expression matrix
}
