% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommPath.R
\name{findLRpath}
\alias{findLRpath}
\title{To find those pathways in which the genesets show overlap with the marker ligand and receptor genes in our dataset}
\usage{
findLRpath(object, category = "all")
}
\arguments{
\item{object}{CommPath object}

\item{category}{Character to indicate which pathway to investigate; one of "go" (GO terms), "kegg" (for KEGG pathways), 'wiki' (for WikiPathways), and "reactome" (for reactome pathways), or "all" for all pathways}
}
\value{
CommPath object containing the ligand-receptor interaction information and the pathways showing overlap with the marker ligand and receptor genes in the dataset
}
\description{
To find those pathways in which the genesets show overlap with the marker ligand and receptor genes in our dataset
}
