% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommPath.R
\name{findReceptor}
\alias{findReceptor}
\title{To find the downstream identity class of specific ligand released by specific upstream identity class}
\usage{
findReceptor(object, select.ident = NULL, select.ligand = NULL, filter = TRUE)
}
\arguments{
\item{object}{CommPath object}

\item{select.ident}{Upstream identity class; if 'NULL', use all identity classes}

\item{select.ligand}{Ligand released by upstream identity class; if 'NULL', use all ligands that are markers for the selected upstream identity class}

\item{filter}{Logical value indicating to find the downstream identity class from the filtered LR interactions or not; default is TRUE}
}
\value{
Data frame including the interaction information
}
\description{
To find the downstream identity class of specific ligand released by specific upstream identity class
}
