% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plug-in.function.R
\name{getPathAttr}
\alias{getPathAttr}
\title{To retrieve the available statistical measures for pathways}
\usage{
getPathAttr(object)
}
\arguments{
\item{object}{CommPath object}
}
\value{
print available statistical measures for pathways
}
\description{
To retrieve the available statistical measures for pathways
}
