% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commpath.R
\name{LRcolor}
\alias{LRcolor}
\title{this is a plug-in function, aiming to get the color for the dots representing receptors and ligands}
\usage{
LRcolor(LRpvalue, user.set.col)
}
\arguments{
\item{LRpvalue}{pvalues of ligands or receptors}

\item{user.set.col}{colors set by user}
}
\value{
Colors matching each p value
}
\description{
this is a plug-in function, aiming to get the color for the dots representing receptors and ligands
}
