% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plug-in.function.R
\name{LRcolor.up.down}
\alias{LRcolor.up.down}
\title{This is a plug-in function, aiming to get the color for the dots representing receptors and ligands}
\usage{
LRcolor.up.down(logfc, user.set.col)
}
\arguments{
\item{logfc}{logfc of ligands or receptors}

\item{user.set.col}{colors set by user}
}
\value{
Colors matching each p value
}
\description{
This is a plug-in function, aiming to get the color for the dots representing receptors and ligands
}
