% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plug-in.function.R
\name{LRinten.to.width}
\alias{LRinten.to.width}
\title{To transform LR intensity to line width}
\usage{
LRinten.to.width(x)
}
\arguments{
\item{x}{LR intensity}
}
\value{
Line width
}
\description{
To transform LR intensity to line width
}
