% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plug-in.function.R
\name{order.and.top}
\alias{order.and.top}
\title{To extract top n element}
\usage{
order.and.top(x, n)
}
\arguments{
\item{x}{vector of elements with names}

\item{n}{top a}
}
\value{
top n elements in x
}
\description{
To extract top n element
}
