% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commpath.R
\name{pasteIdent}
\alias{pasteIdent}
\title{This is a plug-in function, aimming to paste a vector of idents into a ',' separated string}
\usage{
pasteIdent(ident.missed)
}
\arguments{
\item{ident.missed}{Vector of idents}
}
\value{
String with idents pasted
}
\description{
This is a plug-in function, aimming to paste a vector of idents into a ',' separated string
}
