% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commpath.R
\name{pathHeatmap}
\alias{pathHeatmap}
\title{To plot a heatmap of those differentially enriched pathways for each clustet}
\usage{
pathHeatmap(
  gsva.mat,
  ident.label,
  all.path.dat,
  top.n.pathway = 10,
  sort = "p.val.adj",
  col = NULL,
  show.legend = FALSE,
  cell.label.size = NULL,
  cell.label.angle = 0,
  pathway.label.size = NULL,
  scale = TRUE,
  truncation = 1.5
)
}
\arguments{
\item{gsva.mat}{Matrix containing the pathway enrichment sorces, with rows representing pathways and columns representing cells. Pathway scores are usually computed from gsva, or other methods aiming to measure the pathway enrichment in cells}

\item{ident.label}{Vector indicating the identity labels of cells, and the order of labels are required to match order of cells (columns) in the gsva.mat}

\item{all.path.dat}{data.frame of differential enrichment test result from diffAllPath}

\item{top.n.pathway}{Show the heatmap of top n most significant pathways}

\item{sort}{Sort criteria used to select the top n pathways, either p.val.adj, p.val, mean.diff, or median.diff}

\item{col}{Vector of colors used to generate a series of gradient colors to show the enrichment score of pathways; provide a vector containing at least two colors}

\item{show.legend}{Whether to show the legend or not; default is FALSE}

\item{cell.label.size}{Text size of the label of cell types}

\item{cell.label.angle}{Rotation angle of the label of cell types}

\item{pathway.label.size}{Text size of the label of pathways}

\item{scale}{Whether to scale the enrichment sorces matrix among cells or not; default is TRUE}

\item{truncation}{Truncation fold value; scores > (the third quartiles + truncation * interquartile range) and scores < (the first quartiles - truncation * interquartile range) will be adjusted; either a value to indicate the specific truncation value or 'none' to indicate no truncation; default is 1.5}
}
\value{
Heatmap plot showing the top enriched patways in each cluster
}
\description{
To plot a heatmap of those differentially enriched pathways for each clustet
}
