% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Commpath-visualization.R
\name{pathInterPlot.compare}
\alias{pathInterPlot.compare}
\title{To compare the pathway mediated cell-cell communication flow for a specific cluster between two Commpath object}
\usage{
pathInterPlot.compare(
  object.1,
  object.2,
  select.ident,
  diff.marker.dat = NULL,
  diff.path.dat = NULL,
  top.n.path = 5,
  path.order = "p.val.adj",
  p.thre = 0.05,
  top.n.receptor = 10,
  top.n.ligand = 10,
  dot.ident.col = NULL,
  dot.ident.size = 1,
  dot.gene.col = NULL,
  dot.gene.size = 1,
  bar.pathway.col = NULL,
  bar.pathway.width = 10,
  label.text.size = 1,
  label.title.size = 1
)
}
\arguments{
\item{object.1}{Commpath object 1}

\item{object.2}{Commpath object 2}

\item{select.ident}{select.ident}

\item{diff.marker.dat}{diff.marker.dat}

\item{diff.path.dat}{diff.path.dat}

\item{top.n.path}{top n pathways with the smallest adjusted p values to plot}

\item{path.order}{path.order}

\item{p.thre}{threshold for adjust p values; Only pathways with a adjust p valua < p.thre would be considered}

\item{top.n.receptor}{top n receptor with the largest log2FCs to plot}

\item{top.n.ligand}{top n ligand with the largest log2FCs to plot}

\item{dot.ident.col}{color of the dots representing clusters}

\item{dot.ident.size}{size of the dots representing clusters}

\item{dot.gene.col}{color of the dots representing receptors}

\item{dot.gene.size}{size of the dots representing receptors}

\item{bar.pathway.col}{color of the bars reprsenting pathways}

\item{bar.pathway.width}{width of the bars reprsenting pathways}

\item{label.text.size}{text size in the plot}

\item{label.title.size}{text size of the title annotation of the plot}
}
\value{
Network plot showing the significantly differentially activated pathways between the selected clusters of two Commpath object, receptors involved in the pathways, and the upstream clusters which show LR connections with the selected cluster
}
\description{
To compare the pathway mediated cell-cell communication flow for a specific cluster between two Commpath object
}
