% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommPath-visualization.R
\name{pathInterPlot}
\alias{pathInterPlot}
\title{To present the interactions for a selected cluster, including both the upstream and downstream clusters which are connected by specific pathways in the selected cluster}
\usage{
pathInterPlot(
  object,
  select.ident,
  acti.path.dat = NULL,
  top.n.path = 5,
  path.order = "P.val.adj",
  p.thre = 0.05,
  top.n.receptor = 10,
  top.n.ligand = 10,
  dot.ident.col = NULL,
  dot.ident.size = 1,
  dot.gene.col = NULL,
  dot.gene.size = 1,
  bar.pathway.col = NULL,
  label.text.size = 1,
  label.title.size = 1
)
}
\arguments{
\item{object}{CommPath object}

\item{select.ident}{Plot the activated pathways for which cluster or cell type?}

\item{acti.path.dat}{Data frame of differential activation test result from diffAllPath}

\item{top.n.path}{Top n pathways with the smallest adjusted p values to plot}

\item{path.order}{Sort criteria used to select the top n pathways, either 'P.val' or 'P.val.adj', which represent the original and adjusted p values, or 'diff' which represents the mean (in t test) or median (in wilcox test) difference}

\item{p.thre}{Threshold for adjust p values; Only pathways with a adjust p valua < p.thre would be considered}

\item{top.n.receptor}{Top n receptor with the largest log2FCs to plot}

\item{top.n.ligand}{Top n ligand with the largest log2FCs to plot}

\item{dot.ident.col}{Color of the dots representing clusters}

\item{dot.ident.size}{Size of the dots representing clusters}

\item{dot.gene.col}{Color of the dots representing receptors and ligands}

\item{dot.gene.size}{Size of the dots representing receptors and ligands}

\item{bar.pathway.col}{Color of the bars reprsenting pathways}

\item{label.text.size}{Text size in the plot}

\item{label.title.size}{Text size of the title annotation of the plot}
}
\value{
Network plot showing receptors in the selected cluster, the upstream clusters which show L-R connections with the selected cluster, and the significant pathways involved in the receptors
}
\description{
To present the interactions for a selected cluster, including both the upstream and downstream clusters which are connected by specific pathways in the selected cluster
}
