% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommPath-main.R
\name{pathNet.lig.path.stat}
\alias{pathNet.lig.path.stat}
\title{To integrate the statistics of LR interactions and the ligand-associated activated pathways}
\usage{
pathNet.lig.path.stat(Interact, acti.path.filtered.dat)
}
\arguments{
\item{Interact}{Dataframe with statistics of filtered LR interactions}

\item{acti.path.filtered.dat}{Data frame of differential enrichment test result from diffAllPath}
}
\value{
Dataframe with statistics of LR interactions and the ligand-associated activated pathways stored
}
\description{
To integrate the statistics of LR interactions and the ligand-associated activated pathways
}
