% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommPath-visualization.R
\name{pathNetPlot}
\alias{pathNetPlot}
\title{To present a net plot for pathways and upstream or downstream associated LR interactions}
\usage{
pathNetPlot(
  object,
  select.ident,
  plot = "upstream",
  ident.col = NULL,
  vert.size.attr = "degree",
  vert.size.LR = 0.5,
  vert.size.path.adj = 5,
  top.n.path = NULL,
  path.order = "P.val.adj",
  select.path = "all",
  layout = "layout.auto",
  LR.label = "LR",
  pathway.label = TRUE,
  edge.arrow.size = 0.2,
  edge.color = "gray75",
  vertex.label.cex.LR = 0.5,
  vertex.label.cex.path = 0.5,
  vertex.label.color = "black",
  vertex.label.family = "Helvetica",
  vertex.frame.color = "#ffffff",
  node.pie = TRUE,
  return.net = FALSE,
  return.data = FALSE
)
}
\arguments{
\item{object}{CommPath object}

\item{select.ident}{Cluster of interest}

\item{plot}{To show the network of activated pathways and the upstream or downstream LR interactions; default is 'upstream'}

\item{ident.col}{Vector of colors of each cluster; names of the col vector are supposed to be assigned to indicate each color for each cluster}

\item{vert.size.attr}{Which statistical measures should be mapped to the size of vertex; Type "getNetAttr(object)" to retrieve the available statistical measures for pathways}

\item{vert.size.LR}{Size of node representing LR interactions}

\item{vert.size.path.adj}{pseudocount to adjust the size of node representing pathways}

\item{top.n.path}{Top n activated pathways in the selected cluster of interest; if NULL, slelect all activated pathways}

\item{path.order}{Sort criteria used to select the top n pathways, either "P.val" or "P.val.adj", which represent the original and adjusted p values, or "diff" which represents the mean (in t test) or median (in wilcox test) difference; default is "P.val.adj". This parameter would be masked if top.n.path is set as NULL.}

\item{select.path}{Vector of pathways for which users want to present a chain plot; default is all, which means to plot all pathways. The parameters top.n.path and path.order would be masked if this one is set with specific pathways.}

\item{layout}{Network layout; defualt is Fruchterman-Reingold ('layout.fruchterman.reingold'). Refer to the igraph package for more information. Note: an error would be raised if the selected layout do not apply to the graph.}

\item{LR.label}{To show which label on the node representing LR interactions. Available options: LR (to show both the ligand and receptor label), L (to show only the ligand label), R (to show only the receptor label), or NA (no label); default is NA}

\item{pathway.label}{Logical value indicating to display the label of pathways or not; default is TRUE}

\item{edge.arrow.size}{Size of arrow}

\item{edge.color}{Color of edge}

\item{vertex.label.cex.LR}{Font size of LR label}

\item{vertex.label.cex.path}{Font size of pathway label}

\item{vertex.label.color}{color of vertex label}

\item{vertex.label.family}{Font family of the label}

\item{vertex.frame.color}{Color of Node border}

\item{node.pie}{Logical value indicating to show the node representing pathways in pie charts or not; default is TRUE}

\item{return.net}{Logical value indicating to return the igraph object or not; default is FALSE}

\item{return.data}{Logical value indicating to return the link dataframe or not; default is FALSE. This parameter allows users to export the link data in a dataframe, which could be analyzed in other network visualization tools, for example Cytoscape.}
}
\value{
Net plot showing the activated pathways and associated LR interactions
}
\description{
To present a net plot for pathways and upstream or downstream associated LR interactions
}
