% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommPath-visualization.R
\name{pathPlot.compare}
\alias{pathPlot.compare}
\title{To compare differentially activated pathways and the involved receptors between the selected clusters of two CommPath object}
\usage{
pathPlot.compare(
  object.1,
  object.2,
  select.ident,
  up.ident = NULL,
  diff.path.dat = NULL,
  top.n.path = 5,
  path.order = "P.val.adj",
  select.path = "all",
  p.thre = 0.05,
  top.n.receptor = 10,
  dot.ident.col = NULL,
  dot.ident.size = 1,
  dot.gene.col = NULL,
  dot.gene.size = 1,
  bar.pathway.col = NULL,
  label.text.size = 1,
  label.title.size = 1
)
}
\arguments{
\item{object.1}{CommPath object 1}

\item{object.2}{CommPath object 2 for comparison}

\item{select.ident}{Plot the activated pathways for which cluster or cell type?}

\item{up.ident}{Upstream clusters in the communication chain; default is all clusters}

\item{diff.path.dat}{Data frame of defferential activation test result from comparePath; if NULL, comparePath would be run to get diff.path.dat}

\item{top.n.path}{Top n pathways with the smallest adjusted p values to plot}

\item{path.order}{Sort criteria used to select the top n pathways, either 'P.val' or 'P.val.adj', which represent the original and adjusted p values, or 'diff' which represents the mean (in t test) or median (in wilcox test) difference}

\item{select.path}{Vector of pathways for which users want to present a chain plot; default is all, which means to plot all pathways. The parameter top.n.path would be masked if this one is set with specific pathways.}

\item{p.thre}{Threshold for adjust p values; Only pathways with a adjust p valua < p.thre would be considered}

\item{top.n.receptor}{Top n receptor with the largest log2FCs to plot}

\item{dot.ident.col}{Color of the dots representing clusters}

\item{dot.ident.size}{Size of the dots representing clusters}

\item{dot.gene.col}{Color of the dots representing receptors}

\item{dot.gene.size}{Size of the dots representing receptors}

\item{bar.pathway.col}{Color of the bars reprsenting pathways}

\item{label.text.size}{Text size in the plot}

\item{label.title.size}{Text size of the title annotation of the plot}
}
\value{
Network plot showing the significantly differentially activated pathways between the selected clusters of two CommPath object, receptors involved in the pathways, and the upstream clusters which show LR connections with the selected cluster
}
\description{
To compare differentially activated pathways and the involved receptors between the selected clusters of two CommPath object
}
