% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commpath.R
\name{pathPlot}
\alias{pathPlot}
\title{To present the interact and variable pathways in a line plot}
\usage{
pathPlot(
  Interact,
  gsva.mat,
  all.test.dat,
  order = NULL,
  n = 10,
  hide.isol = TRUE,
  ident.size = 1,
  label.size = 1,
  path.size = 1,
  ident.line.width = 1,
  ident.line.alpha = 0.8,
  path.line.alpha = 0.8,
  path.line.width = 1,
  label.dist = 0.2
)
}
\arguments{
\item{Interact}{Interact list returned by findLRpairs}

\item{gsva.mat}{pathway score}

\item{all.test.dat}{all.test.dat}

\item{order}{Order of identity classes in the plot}

\item{n}{Top n variable pathways}

\item{hide.isol}{Whether to hide the isolated identity (those identity class in which cells show no interact with cells in other ientity class) or not; defult is TRUE}

\item{line.width}{line.width}
}
\value{
String with idents pasted
}
\description{
To present the interact and variable pathways in a line plot
}
