% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommPath.R
\name{pathTest}
\alias{pathTest}
\title{Differential enrichment analysis by t.test or wilcox.txt}
\usage{
pathTest(
  acti.ident.score,
  group,
  select.ident.1,
  select.ident.2 = NULL,
  method = "t.test",
  only.posi = FALSE,
  only.sig = FALSE
)
}
\arguments{
\item{acti.ident.score}{Matrix of pathway scores, pathway * cell}

\item{group}{Vector of group labels of cells}

\item{select.ident.1}{Identity class 1}

\item{select.ident.2}{Identity class 2 for comparison}

\item{method}{Method for hypothesis test, either 't.test' or 'wilcox.test'}

\item{only.posi}{only save the information of pathways showing up regulation}

\item{only.sig}{only save the information of pathways showing significant differences}
}
\value{
Data frame including the statistic result
}
\description{
Differential enrichment analysis by t.test or wilcox.txt
}
