% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commpath.R
\name{receptorPathPlot}
\alias{receptorPathPlot}
\title{To present the interact and variable pathways in a line plot, for a specific ident}
\usage{
receptorPathPlot(
  Interact,
  select.ident,
  ident.path.dat,
  top.n.receptor = 5,
  top.n.path = 10,
  order = NULL,
  p.thre = 0.05,
  ident.size = 1,
  receptor.size = 1,
  label.dist = 0.4,
  label.size = 1,
  line.size = 1,
  ident.line.alpha = 1,
  ident.line.width = 1,
  path.line.width = 1,
  path.line.alpha = 0.8,
  path.size = 1
)
}
\arguments{
\item{Interact}{Interact list returned by findLRpairs}

\item{select.ident}{select.ident}

\item{ident.path.dat}{ident.path.dat for the selected ident}

\item{top.n.receptor}{top.n.receptor}

\item{order}{order}

\item{p.thre}{p.thre}

\item{ident.size}{ident.size}
}
\value{
String with idents pasted
}
\description{
To present the interact and variable pathways in a line plot, for a specific ident
}
