% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plug-in.function.R
\name{rotated.axis.element.text}
\alias{rotated.axis.element.text}
\title{To set up the hjust and vjust of text on axises}
\usage{
rotated.axis.element.text(angle, position = "x")
}
\arguments{
\item{angle}{rotation angle}

\item{position}{'x' or 'y' axis}
}
\value{
Theme element_text
}
\description{
To set up the hjust and vjust of text on axises
}
