% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Commpath-object.R
\docType{methods}
\name{show,Commpath-method}
\alias{show,Commpath-method}
\title{show method for Commpath}
\usage{
\S4method{show}{Commpath}(object)
}
\arguments{
\item{object}{A Commpath object}
}
\description{
show method for Commpath
}
