% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plug-in.function.R
\name{subsetMatrix}
\alias{subsetMatrix}
\title{To subset a matrix while keeping the dimension}
\usage{
subsetMatrix(mat, row.select = NULL, col.select = NULL)
}
\arguments{
\item{mat}{matrix}

\item{row.select}{rows slected}

\item{col.select}{columns selected}
}
\value{
subset of the matrix
}
\description{
To subset a matrix while keeping the dimension
}
