% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commpath.R
\name{variPath}
\alias{variPath}
\title{This is a plug-in function, aimming to find the highly variable pathways in the gsva score matrix}
\usage{
variPath(gsva.mat, select.path = NULL, n = 10)
}
\arguments{
\item{gsva.mat}{Matrix containing the pathway enrichment sorces, with rows representing pathways and columns representing cells. Pathway scores are usually computed from gsva, or other methods aiming to measure the pathway enrichment in cells}

\item{select.path}{selected pathways}

\item{n}{n}
}
\value{
Vector of top n variable pathways
}
\description{
This is a plug-in function, aimming to find the highly variable pathways in the gsva score matrix
}
