% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommPath-visualization.R
\name{circosPlot}
\alias{circosPlot}
\title{To present a circos plot}
\usage{
circosPlot(
  object,
  plot = "count",
  ident.col = NULL,
  filter = TRUE,
  select.ident = NULL,
  name.vert = FALSE
)
}
\arguments{
\item{object}{CommPath object}

\item{plot}{To present a circos plot for LR count ("count") or overall interaction intensity ("intensity") among cell clusters}

\item{ident.col}{Vector of colors of each cluster; names of the ident.col vector are supposed to be assigned to indicate each color for each cluster}

\item{filter}{Logical value indicating to present the circos plot for the filtered LR interactions or not; default is TRUE}

\item{select.ident}{To highlight the interaction between a specific cluster class and others; if 'NULL', plot interaction for all clusters}

\item{name.vert}{Should the group annotation be vertical to the grid? Defualt is FALSE}
}
\value{
Circos plot showing the ligand-receptor interaction
}
\description{
To present a circos plot
}
