% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Commpath-visualization.R
\name{circosPlot}
\alias{circosPlot}
\title{To present a circos plot}
\usage{
circosPlot(object, order = NULL, col = NULL, ident = NULL, name.vert = FALSE)
}
\arguments{
\item{object}{Commpath object}

\item{order}{Vector of orders of the identities arround the circos plot}

\item{col}{Vector of colors of each identity; names of the col vector are supposed to be assigned to indicate each color for each identity}

\item{ident}{To highlight the interaction between a specific identity class and others; if 'NULL', plot interaction for all identity classes}

\item{name.vert}{Should the group annotation be vertical to the grid? Defualt is FALSE}
}
\value{
Circos plot showing the ligand-receptor interaction
}
\description{
To present a circos plot
}
