% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommPath-object.R
\docType{class}
\name{CommPath-class}
\alias{CommPath-class}
\alias{CommPath}
\title{To define the CommPath object and the key slots}
\description{
To define the CommPath object and the key slots
}
\section{Slots}{

\describe{
\item{\code{data}}{normalized expression matrix}

\item{\code{cell.info}}{vector of cell labels}

\item{\code{meta.info}}{list of important parameters used during the analysis}

\item{\code{LR.marker}}{data frame of the differential expression test result of ligands and receptors}

\item{\code{interact}}{list containing information of LR interaction among clusters}

\item{\code{interact.filter}}{list containing information of filtered LR interaction among clusters}

\item{\code{pathway}}{list containing information of pathways associated with ligands and receptors}

\item{\code{pathway.net}}{list containing the integrated information of the statistics of LR interactions and associated pathways}
}}

