% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CommpathData}
\alias{CommpathData}
\title{Ligand-receptor pairs and pathway annotations in Commpath database}
\format{
A list includes ligand-receptor pairs and pathway annotations including KEGG pathways, WikiPathways, reactome pathways, and GO terms
}
\usage{
CommpathData
}
\description{
The ligand-receptor pairs and pathway annotations curated in Commpath
}
\keyword{datasets}
