% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommPath-object.R
\name{createCommPath}
\alias{createCommPath}
\title{To create a CommPath object}
\usage{
createCommPath(expr.mat, cell.info, species)
}
\arguments{
\item{expr.mat}{Matrix or data frame of expression matrix, with genes in rows and cells in columns}

\item{cell.info}{Vector of lables indicating identity classes of cells in the expression matrix, and the order of lables should match the order of cells in the expression matrix; or a data frame containing the meta infomation of cells with the row names matching the cells in the expression matrix and a column named as "Cluster" must be included to indicate identity classes of cells}

\item{species}{Species, either 'hsapiens' or 'mmusculus'}
}
\value{
CommPath object
}
\description{
To create a CommPath object
}
