% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Commpath-object.R
\name{createCommpath}
\alias{createCommpath}
\title{To create a Commpath object}
\usage{
createCommpath(expr.mat, cell.info, species)
}
\arguments{
\item{expr.mat}{Matrix or data frame of expression matrix, with genes in rows and cells in columns}

\item{cell.info}{data.frame of mata information of cells, and the row names should match the column names of expr.mat; a column named as "Cluster" should be included in the data.frame and this column would indicate the identity classes of cells in the expression matrix}

\item{species}{Species}
}
\value{
Commpath object
}
\description{
To create a Commpath object
}
