% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Commpath.R
\name{diffAllPath}
\alias{diffAllPath}
\title{To find differentially activated pathways in each identity class}
\usage{
diffAllPath(object, method = "t.test")
}
\arguments{
\item{object}{Commpath object}

\item{method}{Method used for differential enrichment analysis, either 't.test' of 'wilcox.test'}
}
\value{
Data frame including the statistic result comparing the pathway enrichment sorces between cells in each cluster and all other clusters, the significant recetor and ligand in the pathways, and the corresponding up stream identity class and ligand
}
\description{
To find differentially activated pathways in each identity class
}
