% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Commpath.R
\name{diffCommpathPath}
\alias{diffCommpathPath}
\title{To conduct defferential activation test in select.ident between two Commpath objects}
\usage{
diffCommpathPath(
  object.1,
  object.2,
  select.ident,
  method = "wilcox.test",
  p.adjust = "BH",
  min.size = 10,
  only.posi = FALSE,
  only.sig = TRUE,
  ...
)
}
\arguments{
\item{object.1}{Commpath object 1}

\item{object.2}{Commpath object 2 for comparison}

\item{select.ident}{Identity class of interest}

\item{method}{Method used for differential expression test, either 'wilcox.test' or 't.test'}

\item{p.adjust}{Method used for p value correction for multiple differential expression test; see p.adjust function for more information}

\item{min.size}{Minimum size of overlaping genes between candidate pathways and the expression matrix}

\item{only.posi}{only logFC > 0}

\item{only.sig}{only p_val_adj < 0.05}

\item{...}{Extra parameters passed to gsva}
}
\value{
Data frame of differentially activated pathways between the same clusters in two Commpath object
}
\description{
To conduct defferential activation test in select.ident between two Commpath objects
}
