% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commpath.R
\name{findLigand}
\alias{findLigand}
\title{To find the upstream identity classes and ligands of specific receptor expressed by specific downstream identity class}
\usage{
findLigand(Interact, select.ident = NULL, select.receptor = NULL)
}
\arguments{
\item{Interact}{Interact list returned by findLRpairs}

\item{select.ident}{Downstream identity class; if 'NULL', use all identity classes}

\item{select.receptor}{Receptor expressed by downstream identity class; if 'NULL', use all receptors that are markers for the selected downstream identity class}
}
\value{
Dataframe including the interaction information
}
\description{
To find the upstream identity classes and ligands of specific receptor expressed by specific downstream identity class
}
