% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommPath.R
\name{findLRmarker}
\alias{findLRmarker}
\title{To identify marker ligands and marker receptors in the expression matrix}
\usage{
findLRmarker(object, method = "wilcox.test", p.adjust = "BH")
}
\arguments{
\item{object}{CommPath object}

\item{method}{Method used for differential expression test, either 'wilcox.test' or 't.test'}

\item{p.adjust}{Method used for p value correction for multiple differential expression test; see p.adjust function for more information}
}
\value{
Data frame containing the differential expression test
}
\description{
To identify marker ligands and marker receptors in the expression matrix
}
