% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Commpath.R
\name{findReceptor}
\alias{findReceptor}
\title{To find the downstream identity class of specific ligand released by specific upstream identity class}
\usage{
findReceptor(object, select.ident = NULL, select.ligand = NULL)
}
\arguments{
\item{object}{Commpath object}

\item{select.ident}{Upstream identity class; if 'NULL', use all identity classes}

\item{select.ligand}{Ligand released by upstream identity class; if 'NULL', use all ligands that are markers for the selected upstream identity class}
}
\value{
Dataframe including the interaction information
}
\description{
To find the downstream identity class of specific ligand released by specific upstream identity class
}
