% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commpath.R
\name{findReceptor}
\alias{findReceptor}
\title{To find the downstream identity class of specific ligand released by specific upstream identity class}
\usage{
findReceptor(Interact, select.ident = NULL, select.receptor = NULL)

findReceptor(Interact, select.ident = NULL, select.receptor = NULL)
}
\arguments{
\item{Interact}{Interact list returned by findLRpairs}

\item{select.ident}{Downstream identity class; if 'NULL', use all identity classes}

\item{select.receptor}{Receptor expressed by downstream identity class; if 'NULL', use all receptors that are markers for the selected downstream identity class}

\item{select.ligand}{Ligand released by upstream identity class; if 'NULL', use all ligands that are markers for the selected upstream identity class}
}
\value{
Dataframe including the interaction information

Dataframe including the interaction information
}
\description{
To find the downstream identity class of specific ligand released by specific upstream identity class

To find the upstream identity classes and ligands of specific receptor expressed by specific downstream identity class
}
