% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commpath.R
\name{orderCheck}
\alias{orderCheck}
\title{This is a plug-in function, aimming to check whether the specified order of idents by users contain all idents that present in the dataset}
\usage{
orderCheck(all.ident, order)
}
\arguments{
\item{all.ident}{Vector of all idents that present in the dataset}

\item{order}{Vector of the specified order of idents by users}
}
\value{
if all idents are not contained, stop the procedure
}
\description{
This is a plug-in function, aimming to check whether the specified order of idents by users contain all idents that present in the dataset
}
