% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commpath.R
\name{p.remove.inf}
\alias{p.remove.inf}
\title{this is a plug-in function, aiming to adjust those p values==0}
\usage{
p.remove.inf(pvalues, scale = 0.1)
}
\arguments{
\item{pvalues}{pvalues}

\item{scale}{scale}
}
\value{
p values with inf adjusted
}
\description{
this is a plug-in function, aiming to adjust those p values==0
}
