% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plug-in.function.R
\name{pasteIdent}
\alias{pasteIdent}
\title{To paste a vector of idents into a ',' separated string}
\usage{
pasteIdent(ident.missed)
}
\arguments{
\item{ident.missed}{Vector of idents}
}
\value{
String with idents pasted
}
\description{
To paste a vector of idents into a ',' separated string
}
