% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Commpath-visualization.R
\name{pathHeatmap}
\alias{pathHeatmap}
\title{To plot a heatmap of those differentially enriched pathways for each cluster}
\usage{
pathHeatmap(
  object,
  acti.path.dat = NULL,
  top.n.pathway = 10,
  path.order = "p.val.adj",
  col = NULL,
  show.legend = FALSE,
  cell.label.size = NULL,
  cell.label.angle = 0,
  pathway.label.size = NULL,
  scale = TRUE,
  truncation = 1.5
)
}
\arguments{
\item{object}{Commpath object}

\item{acti.path.dat}{data.frame of differential enrichment test result from diffAllPath; if NULL, diffAllPath would be run to get the acti.path.dat}

\item{top.n.pathway}{Show the heatmap of top n most significant pathways}

\item{path.order}{Sort criteria used to select the top n pathways, either 'p.val' or 'p.val.adj', which represent the original and adjusted p values, or 'diff' which represents the mean (in t test) or median (in wilcox test) difference}

\item{col}{Vector of colors used to generate a series of gradient colors to show the enrichment score of pathways; provide a vector containing at least two colors}

\item{show.legend}{Whether to show the legend or not; default is FALSE}

\item{cell.label.size}{Text size of the label of cell types}

\item{cell.label.angle}{Rotation angle of the label of cell types}

\item{pathway.label.size}{Text size of the label of pathways}

\item{scale}{Whether to scale the enrichment sorces matrix among cells or not; default is TRUE}

\item{truncation}{Truncation fold value; scores > (the third quartiles + truncation * interquartile range) and scores < (the first quartiles - truncation * interquartile range) will be adjusted; either a value to indicate the specific truncation value or 'none' to indicate no truncation; default is 1.5}
}
\value{
Heatmap plot showing the top enriched patways in each cluster
}
\description{
To plot a heatmap of those differentially enriched pathways for each cluster
}
