% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommPath.R
\name{pathNet}
\alias{pathNet}
\title{To integrate the statistics of LR interactions and associated activated pathways}
\usage{
pathNet(object, acti.path.filtered.dat)
}
\arguments{
\item{object}{CommPath object}

\item{acti.path.filtered.dat}{Data frame of differential enrichment test result from diffAllPath}
}
\value{
CommPath object with statistics of LR interactions and associated pathways saved in the slot pathway.net
}
\description{
To integrate the statistics of LR interactions and associated activated pathways
}
