% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Commpath-visualization.R
\name{pathPlot}
\alias{pathPlot}
\title{To present the interactions for a selected cluster, including the upstream clusters and the activated pathways in the selected cluster}
\usage{
pathPlot(
  object,
  select.ident,
  acti.path.dat = NULL,
  top.n.path = 5,
  path.order = "p.val.adj",
  p.thre = 0.05,
  top.n.receptor = 10,
  dot.ident.col = NULL,
  dot.ident.size = 1,
  dot.gene.col = NULL,
  dot.gene.size = 1,
  bar.pathway.col = NULL,
  bar.pathway.width = 10,
  label.text.size = 1,
  label.title.size = 1
)
}
\arguments{
\item{object}{Commpath object}

\item{select.ident}{select.ident}

\item{acti.path.dat}{acti.path.dat}

\item{top.n.path}{top n pathways with the smallest adjusted p values to plot}

\item{path.order}{Sort criteria used to select the top n pathways, either 'p.val' or 'p.val.adj', which represent the original and adjusted p values, or 'diff' which represents the mean (in t test) or median (in wilcox test) difference}

\item{p.thre}{threshold for adjust p values; Only pathways with a adjust p valua < p.thre would be considered}

\item{top.n.receptor}{top n receptor with the largest log2FCs to plot}

\item{dot.ident.col}{color of the dots representing clusters}

\item{dot.ident.size}{size of the dots representing clusters}

\item{dot.gene.col}{color of the dots representing receptors}

\item{dot.gene.size}{size of the dots representing receptors}

\item{bar.pathway.col}{color of the bars reprsenting pathways}

\item{bar.pathway.width}{width of the bars reprsenting pathways}

\item{label.text.size}{text size in the plot}

\item{label.title.size}{text size of the title annotation of the plot}
}
\value{
Network plot showing the significantly activated pathways in the select cluster compared to all other clusters, receptors involved in the pathways, and the upstream clusters which show LR connections with the selected cluster
}
\description{
To present the interactions for a selected cluster, including the upstream clusters and the activated pathways in the selected cluster
}
