% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commpath.R
\name{pathTest}
\alias{pathTest}
\title{Differential enrichment analysis by t.test or wilcox.txt}
\usage{
pathTest(
  gsva.ident.mat,
  group,
  select.ident.1,
  select.ident.2 = NULL,
  method = "t.test"
)
}
\arguments{
\item{gsva.ident.mat}{Matrix of pathway scores, pathway * cell}

\item{group}{Vector of group labels of cells}

\item{select.ident.1}{Identity class 1}

\item{select.ident.2}{Identity class 2 for comparison}

\item{method}{Method for hypothesis test, either 't.test' or 'wilcox.test'}
}
\value{
Dataframe including the statistic result
}
\description{
Differential enrichment analysis by t.test or wilcox.txt
}
