% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Commpath-visualization.R
\name{receptorPathPlot}
\alias{receptorPathPlot}
\title{To present the interact and variable pathways in a line plot, for a specific ident}
\usage{
receptorPathPlot(
  object,
  select.ident,
  ident.path.dat = NULL,
  path.order = "p.val.adj",
  top.n.receptor = 5,
  order = NULL,
  top.n.path = 10,
  p.thre = 0.05,
  dot.ident.col = NULL,
  dot.receptor.col = NULL,
  bar.pathway.col = NULL,
  bar.pathway.width = 10,
  dot.ident.size = 1,
  dot.receptor.size = 1,
  label.text.size = 1,
  label.title.size = 1,
  line.ident.width = 1,
  line.path.width = 1
)
}
\arguments{
\item{object}{Commpath object}

\item{select.ident}{select.ident}

\item{ident.path.dat}{ident.path.dat for the selected ident; if NULL, diffPath would be run to get the ident.path.dat}

\item{top.n.receptor}{top n receptor with the largest log2FCs to plot}

\item{order}{order of source clusters}

\item{top.n.path}{top n pathways with the smallest adjusted p values to plot}

\item{p.thre}{threshold for adjust p values; Only pathways with a adjust p valua < p.thre would be considered}

\item{dot.ident.col}{color of the dots representing upstream clusters}

\item{dot.receptor.col}{color of the dots representing reptors}

\item{bar.pathway.col}{color of the bars reprsenting pathways}

\item{bar.pathway.width}{width of the bars reprsenting pathways}

\item{dot.ident.size}{size of the dots representing source clusters}

\item{dot.receptor.size}{size of the dots representing receptors}

\item{label.text.size}{text size in the plot}

\item{label.title.size}{text size of the title annotation of the plot}

\item{line.ident.width}{text size of pathway labels}

\item{line.path.width}{width of lines connecting receptors and pathways}
}
\value{
Network plot showing receptors in the selected cluster, the upstream clusters which show L-R connections with the selected cluster, and the significant pathways involved in the receptors
}
\description{
To present the interact and variable pathways in a line plot, for a specific ident
}
