% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plug-in.function.R
\name{scale_1}
\alias{scale_1}
\title{To scale x with minimum equal to 1 and maximum equal to 2}
\usage{
scale_1(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
scaled x with minimum equal to 1 and maximum equal to 2
}
\description{
To scale x with minimum equal to 1 and maximum equal to 2
}
