% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommPath.R
\name{scorePath}
\alias{scorePath}
\title{To find different enriched pathway between two group cells}
\usage{
scorePath(object, method = "gsva", min.size = 10, ...)
}
\arguments{
\item{object}{CommPath object}

\item{method}{Method used for scoring the pathways, either 'gsva' of 'average'}

\item{min.size}{Minimum size of overlaping genes between candidate pathways and the expression matrix}

\item{...}{Extra parameters passed to gsva}
}
\value{
CommPath object with pathways activation scores stored in the slot pathway
}
\description{
To find different enriched pathway between two group cells
}
