% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommPath-object.R
\docType{methods}
\name{show,CommPath-method}
\alias{show,CommPath-method}
\title{Show method for CommPath}
\usage{
\S4method{show}{CommPath}(object)
}
\arguments{
\item{object}{A CommPath object}
}
\description{
Show method for CommPath
}
