% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plug-in.function.R
\name{subsetCommpath}
\alias{subsetCommpath}
\title{To subset a Commpath object}
\usage{
subsetCommpath(object, ident.keep)
}
\arguments{
\item{object}{Commpath object}

\item{ident.keep}{idents to keep}
}
\value{
a subset Commpath object
}
\description{
To subset a Commpath object
}
