% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plug-in.function.R
\name{subsetCommPath}
\alias{subsetCommPath}
\title{To subset a CommPath object}
\usage{
subsetCommPath(object, ident.keep)
}
\arguments{
\item{object}{CommPath object}

\item{ident.keep}{idents to keep}
}
\value{
a subset CommPath object
}
\description{
To subset a CommPath object
}
